requirements.txt
transformers>=4.44.0
torch>=2.1.0
torchaudio>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
tk
